Refactor button styles
Move most SmallButton styling as global styles, and its flexbox styling as a composition style. Support global styling of 'a' elements as buttons via `:where(.button)` to keep low specificity.

Have the complexity of BigButton styling in block styling, and the color variation as exception styling. Rename the prop for color to be clear on what it controls.