Use const tag for dynamic component expression
The mapping of results to objects holding all values was only made due to the limitation of not being able to use an object indexing expression as a dynamic component tag name. Having an array remapping feels like excessive work for the benefit of one field.
With the const tag, the expression can be saved to a variable name compatible with dynamic component rendering.