Override cookie version installed for SvelteKit
Months ago there was an attempt to update a dependency on a vulnerable version of cookie, which was reverted due to the change needed technically having a breaking change behavior for cookie naming. The Svelte team did update the compiler to warn about invalid cookie names, but plan to update the cookie dependency in SvelteKit v3.

Overriding to install a non-vulnerable version of cookie has been encouraged for now.